tests: net_pkt: Make testcase looks similar to other testcases
All test case should have entry function in main.c file and
other functions should be called from separate file.

JIRA: ZEP-2249
