+++ tests/net/net_pkt/src/Makefile
@@ -1,4 +1,4 @@
-obj-y = main.o
+obj-y = main.o test_net_pkt_context.o
 ccflags-y += -I${ZEPHYR_BASE}/subsys/net/ip
 
 include $(ZEPHYR_BASE)/tests/Makefile.test
